requirements.txt
aiogram==3.4.1
apscheduler==3.10.4
python-dotenv==1.0.0